ServiceWA/requirements.txt
fileutilities==1.0.0
robotframework==6.0.2
robotframework-appiumlibrary==1.6.2
robotframework-seleniumlibrary==5.1.3
robotframework-assertion-engine==1.0.0
robotframework-browser==16.0.0
robotframework-pythonlibcore==4.1.2
xlrd
openpyxl
robotframework-datadriver[XLS]==1.4.1
robotframework-pabot==2.0.0
robotframework-excellib==2.0.1
jproperties==2.1.0
XlsxWriter==3.0.1
pytesseract
perfecto==3.0.19
perfecto-py37==0.1.0
perfectolibrary==0.4.9
perfectolibrary-py37==0.5.1
requests
robotframework-requests